requirements.txt
astroid==2.4.2
autopep8==1.5.5
beautifulsoup4==4.9.1
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.5
docutils==0.16
idna==2.10
isort==5.4.2
jeepney==0.6.0
keyring==22.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.9
pkginfo==1.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.3.1
six==1.15.0
soupsieve==2.0.1
spotipy==2.16.0
toml==0.10.1
tqdm==4.56.2
twine==3.3.0
urllib3==1.26.3
vk-api==11.8.0
webencodings==0.5.1
wrapt==1.12.1
yandex-music==1.0.0
